Enable to search for multiple separate keywords
Do not force a search with query "foo bar" to have name exactly
like "foo bar", but search for both "foo" AND "bar" separately.
This allows you to search for recipes named "bar foo" as well.

Also do not return a server error if no query was given, so
return zero results for an empty query string.