Add form to create new recipe.
In doing this I updated the page that displays the recipe so that it no longer
assumes that every recipe has one and only one category, but injects the 'none'
category.

I've also failed to include several other aspects of a recipe in the initial
creation - tags, image, etc. I expect to add these to a recipe edit page later.